Create README.md
This project is based on a large no of tweets posted on Twitter. Here I tried to extract some information from the tweets using some NLP functions, At first, I have cleaned the text by removing punctuations, stopwords, and some noise from the text. Then started to perform some actions such as keyword analysis which is which are the major keywords present in the text. Top mentions which are basically finding the topmost mentions in the text (@), top hashtags, top URLs, NER, and sentiment analysis.